Rename utils.py into tools.py to avoid duplicated name. [TODO] fix this bug in pre-commit
